+++ CHANGELOG.md
@@ -7,11 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- BREAKING: Update `reqwest` to v0.12.x which is based on `hyper`, `http` and `http-body` v1
+
 ## [0.2.0] - 2023-10-14
 
 ### Changed
 
-- BREAKING: use `i64`for timestamps
+- BREAKING: use `i64` for timestamps
 - Make timestamp parsing more resilient by accepting both strings and integers
 - BREAKING: Deserialize maybe-empty strings as `Option<String>`