+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.4.0] - 2024-07-22
+
 ### Changed
 
 - BREAKING: change `duration` from `Duration` to `Option<Duration>` because some entries (e.g. livestreams) don't have a duration 
@@ -50,7 +52,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial release
 
-[Unreleased]: https://github.com/d-k-bo/mediathekviewweb-rs/compare/v0.3.1...HEAD
+[Unreleased]: https://github.com/d-k-bo/mediathekviewweb-rs/compare/v0.4.0...HEAD
+[0.4.0]: https://github.com/d-k-bo/mediathekviewweb-rs/compare/v0.3.1...v0.4.0
 [0.3.1]: https://github.com/d-k-bo/mediathekviewweb-rs/compare/v0.3.0...v0.3.1
 [0.3.0]: https://github.com/d-k-bo/mediathekviewweb-rs/compare/v0.2.0...v0.3.0
 [0.2.0]: https://github.com/d-k-bo/mediathekviewweb-rs/compare/v0.1.2...v0.2.0